# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Insect Monitoring API",
    "description": "곤충 모니터링 API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/"
    }
  ],
  "paths": {
...
    "/collection/list": {
      "get": {
        "tags": [
          "collection"
        ],
        "summary": "채집 이력 목록 (관리자)",
        "description": "",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "count",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionHistoryListResponse"
                }
              }
            }
          },
          "400": {
            "description": "실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralFailedResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/collection/map": {
      "get": {
        "tags": [
          "collection"
        ],
        "summary": "지도용 채집 이력",
        "description": "",
        "parameters": [
          {
            "name": "address_sido",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "address_gungu",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "address_dong",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionHistoryListResponse"
                }
              }
            }
          },
          "400": {
            "description": "실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralFailedResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/collection/detail": {
      "get": {
        "tags": [
          "collection"
        ],
        "summary": "채집 이력 상세",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionHistory"
                }
              }
            }
          },
          "400": {
            "description": "실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralFailedResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/collection/register": {
      "post": {
        "tags": [
          "collection"
        ],
        "summary": "채집 이력 등록",
        "description": "",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionHistory"
                }
              }
            }
          },
          "400": {
            "description": "실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralFailedResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "insect_name",
                  "address_sido",
                  "address_gungu",
                  "address_dong",
                  "latitude",
                  "longitude",
                  "collect_count",
                  "status"
                ],
                "properties": {
                  "device_id": {
                    "type": "integer",
                    "nullable": true
                  },
                  "insect_name": {
                    "type": "string"
                  },
                  "image_file": {
                    "type": "string"
                  },
                  "address_sido": {
                    "type": "string"
                  },
                  "address_gungu": {
                    "type": "string"
                  },
                  "address_dong": {
                    "type": "string"
                  },
                  "address_detail": {
                    "type": "string"
                  },
                  "latitude": {
                    "type": "number"
                  },
                  "longitude": {
                    "type": "number"
                  },
                  "collect_count": {
                    "type": "integer"
                  },
                  "status": {
                    "type": "string",
                    "description": "good | normal | warning | bad"
                  },
                  "memo": {
                    "type": "string"
                  },
                  "created_date": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/collection/update": {
      "put": {
        "tags": [
          "collection"
        ],
        "summary": "채집 이력 수정 (관리자)",
        "description": "",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionHistory"
                }
              }
            }
          },
          "400": {
            "description": "실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralFailedResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id"
                ],
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "device_id": {
                    "type": "integer",
                    "nullable": true
                  },
                  "insect_name": {
                    "type": "string"
                  },
                  "image_file": {
                    "type": "string"
                  },
                  "address_sido": {
                    "type": "string"
                  },
                  "address_gungu": {
                    "type": "string"
                  },
                  "address_dong": {
                    "type": "string"
                  },
                  "address_detail": {
                    "type": "string"
                  },
                  "latitude": {
                    "type": "number"
                  },
                  "longitude": {
                    "type": "number"
                  },
                  "collect_count": {
                    "type": "integer"
                  },
                  "status": {
                    "type": "string"
                  },
                  "memo": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/collection/delete": {
      "delete": {
        "tags": [
          "collection"
        ],
        "summary": "채집 이력 삭제 (관리자)",
        "description": "",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralFailedResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id"
                ],
                "properties": {
                  "id": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CollectionHistory": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  }
                }
              },
              "device_id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "nullable": {
                    "type": "boolean",
                    "example": true
                  }
                }
              },
              "insect_name": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "image_file": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "address_sido": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "address_gungu": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "address_dong": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "address_detail": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "latitude": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "number"
                  }
                }
              },
              "longitude": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "number"
                  }
                }
              },
              "collect_count": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  }
                }
              },
              "status": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "good | normal | warning | bad"
                  }
                }
              },
              "memo": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "created_date": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date-time"
                  }
                }
              }
            }
          }
        }
      },
      "CollectionHistoryListResponse": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "array"
          },
          "items": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "object"
              },
              "properties": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "integer"
                      }
                    }
                  },
                  "device_id": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "integer"
                      },
                      "nullable": {
                        "type": "boolean",
                        "example": true
                      }
                    }
                  },
                  "insect_name": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "string"
                      }
                    }
                  },
                  "image_file": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "string"
                      }
                    }
                  },
                  "address_sido": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "string"
                      }
                    }
                  },
                  "address_gungu": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "string"
                      }
                    }
                  },
                  "address_dong": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "string"
                      }
                    }
                  },
                  "address_detail": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "string"
                      }
                    }
                  },
                  "latitude": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "number"
                      }
                    }
                  },
                  "longitude": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "number"
                      }
                    }
                  },
                  "collect_count": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "integer"
                      }
                    }
                  },
                  "status": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "string"
                      }
                    }
                  },
                  "memo": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "string"
                      }
                    }
                  },
                  "created_date": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "string"
                      },
                      "format": {
                        "type": "string",
                        "example": "date-time"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "GeneralSuccessResponse": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "type": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "message": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              }
            }
          }
        }
      },
      "GeneralFailedResponse": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "type": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "message": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}